moves the types to one file
