+++ src/signals/incident/components/form/ContainerSelect/ContainerLayer/index.ts
@@ -1,0 +1,3 @@
+import ContainerLayer from './ContainerLayer';
+
+export default ContainerLayer;
