+++ src/signals/incident/components/form/ContainerSelect/WfsLayer/index.ts
@@ -1,0 +1,3 @@
+import WfsLayer from './WfsLayer';
+
+export default WfsLayer;
