+++ src/signals/incident/components/form/ContainerSelect/types.ts
@@ -1,4 +1,5 @@
-import type { IconOptions, LatLngExpression } from 'leaflet';
+import type { IconOptions, LatLngExpression, Map } from 'leaflet';
+import type { ZoomLevel } from '@amsterdam/arm-core/lib/types';
 
 export type ClickEvent = (e: React.MouseEvent<HTMLButtonElement | HTMLAnchorElement>) => void;
 
@@ -41,4 +42,16 @@ export interface ContainerSelectValue {
   update: (items: Item[] | null) => void;
   edit: () => void;
   close: () => void;
+}
+
+export interface WfsLayerProps {
+  url: string;
+  options: {
+    getBBox: (mapInstance: Map) => string;
+  };
+  zoomLevel: ZoomLevel;
+}
+
+export interface DataLayerProps {
+  featureTypes: FeatureType[];
 }
